Moved and refactored most tests as common portable versions
